refactor:add cpu_no_boost_1_6ghz cmdline when needed
Changelogs:
Some chips, during ATE testing, have demonstrated that the CPU can run
at 1.6GHz with a voltage of 0.8V. When this information is read from the
OTP, U-Boot will add the parameter cpu_no_boost_1_6ghz to the cmdline.
During Linux startup, this parameter will be checked. For this type of
chip, when Linux boosts the CPU to 1.6GHz, it will not increase the
voltage, thereby reducing power consumption. It is important to note,
however, that when the CPU runs at 1.8GHz, a voltage increase is still
required.

Signed-off-by: huangyifeng <huangyifeng@eswincomputing.com>